Remove head stuff from Scala html
