+++ js/src/main/scala/vyxal/Main.scala
@@ -11,24 +11,20 @@ object Main {
       event => {
         val kb = document.getElementById("keyboard")
         for (i <- codepage.indices) {
-          kb.innerHTML += (`<span class=\"key\" style="text-align:center;" title='`
-            + repr(codepage_descriptions[i])
-            + `'>${codepage[i]}</span>`)
+          kb.innerHTML +=
+            s"""<span class=\"key\" style="text-align:center;"
+                  title='${repr(codepage_descriptions[i])}'>${codepage[i]}</span>""""
         }
         for (item <- document.querySelectorAll(".key")) {
           item.addEventListener(
             "click",
             event => {
               val char = replaceHTMLChar(event.target.innerHTML)
-              val cm = globalThis(s"e_$selectedBox")
-              cm.replaceSelection(char)
-              cm.save()
-              cm.focus()
               JSVyxal.updateCount()
             }
           )
         }
-        og_keyboard_html = document.getElementById("keyboard").innerHTML
+        JSVyxal.og_keyboard_html = document.getElementById("keyboard").innerHTML
       }
     )
 
@@ -44,7 +40,41 @@ object Main {
         val filter = document.getElementById("filterBox")
 
         def do_run() = {
-          if (!run.innerHTML.includes("fa-spin")) {
+          
+        }
+
+        document.getElementById("clear").onclick = 
+          e => {
+            e_code.doc.setValue("")
+            stdin.value = ""
+            output.value = ""
+            extra.value = ""
+            e_footer.doc.setValue("")
+            e_header.doc.setValue("")
+            updateCount()
+            flags.value = ""
+            filter.value = ""
+            glyphSearch()
+            expandBoxes()
+          }
+        
+
+      }
+    )
+
+    document.addEventListener(
+      "keydown",
+      event => {
+        if (event.metaKey && event.key == "Enter") {
+          Body.runButton.click()
+        }
+      }
+    )
+
+  }
+
+  def run() = {
+    if (!run.innerHTML.includes("fa-spin")) {
             run.innerHTML =
               """<svg class="fa-spin" style="width:24px;height:24px" viewBox="0 0 24 24">
                     <path fill="currentColor" d="M12,4V2A10,10 0 0,0 2,12H4A8,8 0 0,1 12,4Z"/>
@@ -66,38 +96,5 @@ object Main {
           } else {
             // todo kill if already running
           }
-        }
-
-        $("#run_button").on("click", do_run)
-
-        $("#clear").on(
-          "click",
-          e => {
-            e_code.doc.setValue("")
-            stdin.value = ""
-            output.value = ""
-            extra.value = ""
-            e_footer.doc.setValue("")
-            e_header.doc.setValue("")
-            updateCount()
-            flags.value = ""
-            filter.value = ""
-            glyphSearch()
-            expandBoxes()
-          }
-        )
-
-      }
-    )
-
-    document.addEventListener(
-      "keydown",
-      (event) => {
-        if (event.metaKey && event.key == "Enter") {
-          $("#run_button").click()
-        }
-      }
-    )
-
   }
 }
